+++ project/Dependencies.scala
@@ -8,6 +8,7 @@ object Dependencies {
     mesos % "compile",
     guava % "compile",
     wixAccord % "compile",
+    scalaLogging % "compile",
     scalaxml % "provided" // for scapegoat
   )
 
@@ -50,6 +51,7 @@ object Dependencies {
     akkaHttpPlayJson % "compile",
     alpakkaS3 % "compile",
     commonsCompress % "compile", // used for tar flow
+    akkaSse % "compile",
 
 
     // test
@@ -61,7 +63,6 @@ object Dependencies {
     Test.scalacheck % "test",
     Test.wixAccordScalatest % "test",
     Test.curatorTest % "test",
-    Test.akkaSse % "test",
     Test.commonsIO % "test"
   ) ++ Kamon.all).map(
     _.excludeAll(excludeSlf4jLog4j12)
@@ -78,45 +79,47 @@ object Dependencies {
 object Dependency {
   object V {
     // runtime deps versions
-    val Aws = "1.11.123"
-    val Alpakka  = "0.7"
+    val Aws = "1.11.128"
+    val Alpakka  = "0.8"
     val Chaos = "0.8.8"
     val Guava = "19.0"
     val Mesos = "1.2.0"
     // Version of Mesos to use in Dockerfile.
     val MesosDebian = "1.2.0-2.0.6"
     val OpenJDK = "openjdk:8u121-jdk"
-    val Akka = "2.4.17"
+    val Akka = "2.4.18"
+    val AkkaHttp = "10.0.6"
+    val AkkaSSE = "2.0.0"
     val ApacheCommonsCompress = "1.13"
     val ApacheCommonsIO = "2.4"
     val AsyncAwait = "0.9.6"
-    val Jersey = "1.18.5"
+    val Jersey = "1.18.6"
     val JettyServlets = "9.3.6.v20151106"
-    val JodaTime = "2.9.7"
+    val JodaTime = "2.9.9"
     val JodaConvert = "1.8.1"
     val UUIDGenerator = "3.1.4"
     val JGraphT = "0.9.3"
     val Diffson = "2.0.2"
-    val PlayJson = "2.5.12"
+    val PlayJson = "2.5.14"
     val JsonSchemaValidator = "2.2.6"
     val RxScala = "0.26.5"
-    val MarathonUI = "1.1.7"
-    val MarathonApiConsole = "3.0.8"
+    val MarathonUI = "1.2.0"
+    val MarathonApiConsole = "3.0.8-accept"
     val Logback = "1.1.3"
-    val Logstash = "4.8"
+    val Logstash = "4.9"
     val WixAccord = "0.5"
     val Curator = "2.11.1"
     val Java8Compat = "0.8.0"
     val ScalaLogging = "3.5.0"
-    val Raven = "7.8.1"
+    val Raven = "7.8.6"
 
     // test deps versions
     val Mockito = "1.10.19"
-    val ScalaTest = "3.0.1"
+    val ScalaTest = "3.0.3"
     val JUnit = "4.12"
     val JUnitBenchmarks = "0.7.2"
     val JMH = "1.14"
-    val ScalaCheck = "1.13.4"
+    val ScalaCheck = "1.13.5"
   }
 
   val excludeMortbayJetty = ExclusionRule(organization = "org.mortbay.jetty")
@@ -127,7 +130,7 @@ object Dependency {
   val akkaActor = "com.typesafe.akka" %% "akka-actor" % V.Akka
   val akkaSlf4j = "com.typesafe.akka" %% "akka-slf4j" % V.Akka
   val akkaStream = "com.typesafe.akka" %% "akka-stream" % V.Akka
-  val akkaHttp = "com.typesafe.akka" %% "akka-http" % "10.0.5"
+  val akkaHttp = "com.typesafe.akka" %% "akka-http" % V.AkkaHttp
   val akkaHttpPlayJson = "de.heikoseeberger" %% "akka-http-play-json" % "1.10.1"
   val asyncAwait = "org.scala-lang.modules" %% "scala-async" % V.AsyncAwait
   val playJson = "com.typesafe.play" %% "play-json" % V.PlayJson
@@ -156,6 +159,7 @@ object Dependency {
   val scalaxml = "org.scala-lang.modules" %% "scala-xml" % "1.0.5"
   val raven = "com.getsentry.raven" % "raven-logback" % V.Raven
   val commonsCompress = "org.apache.commons" % "commons-compress" % V.ApacheCommonsCompress
+  val akkaSse = "de.heikoseeberger" %% "akka-sse" % V.AkkaSSE
 
   object Kamon {
     val Version = "0.6.5"
@@ -187,7 +191,6 @@ object Dependency {
     val scalacheck = "org.scalacheck" %% "scalacheck" % V.ScalaCheck
     val wixAccordScalatest = "com.wix" %% "accord-scalatest" % V.WixAccord
     val curatorTest = "org.apache.curator" % "curator-test" % V.Curator
-    val akkaSse = "de.heikoseeberger" %% "akka-sse" % "2.0.0"
     val commonsIO = "commons-io" % "commons-io" % V.ApacheCommonsIO
   }
 }