+++ src/main/scala/mesosphere/marathon/api/CORSFilter.scala
@@ -9,11 +9,7 @@ import mesosphere.marathon.stream.Implicits._
 
 class CORSFilter @Inject() (config: MarathonConf) extends Filter {
 
-  // Map access_control_allow_origin flag into separate headers
-  lazy val maybeOrigins: Option[Seq[String]] =
-    config.accessControlAllowOrigin.get.map { configValue =>
-      configValue.split(",").map(_.trim)(collection.breakOut)
-    }
+  lazy val maybeOrigins: Option[Seq[String]] = config.accessControlAllowOrigin.get
 
   override def init(filterConfig: FilterConfig): Unit = {}