+++ src/test/scala/mesosphere/marathon/api/MarathonExceptionMapperTest.scala
@@ -49,7 +49,7 @@ class MarathonExceptionMapperTest extends UnitTest {
       val entityString = response.getEntity.asInstanceOf[String]
       val entity = Json.parse(entityString)
       (entity \ "message").as[String] should be("Invalid JSON")
-      (entity \ "details").as[String] should be("""Unexpected end-of-input: expected close marker for OBJECT (from [Source: {"id":"/test"; line: 1, column: 1])""")
+      (entity \ "details").as[String] should be("""Unexpected end-of-input: expected close marker for Object (start marker at [Source: {"id":"/test"; line: 1, column: 1])""")
     }
 
     "Render json mapping exception correctly" in {